+++ admin/import.js
@@ -1,20 +1,20 @@
 /**
  * Import given settings file via AJAX.
  */
-function efml_import_settings_file() {
+function settings_import_file() {
   let file = jQuery('#import_settings_file')[0].files[0];
   if( undefined === file ) {
     let dialog_config = {
       detail: {
-        title: efmlImportJsVars.title_settings_import_file_missing,
+        title: settingsImportJsVars.title_settings_import_file_missing,
         texts: [
-          '<p>' + efmlImportJsVars.text_settings_import_file_missing + '</p>'
+          '<p>' + settingsImportJsVars.text_settings_import_file_missing + '</p>'
         ],
         buttons: [
           {
             'action': 'closeDialog();',
             'variant': 'primary',
-            'text': efmlImportJsVars.lbl_ok
+            'text': settingsImportJsVars.lbl_ok
           }
         ]
       }
@@ -25,11 +25,11 @@ function efml_import_settings_file() {
 
   let request = new FormData();
   request.append( 'file', file);
-  request.append( 'action', 'eml_settings_import_file' );
-  request.append( 'nonce', efmlImportJsVars.settings_import_file_nonce );
+  request.append( 'action', 'settings_import_file' );
+  request.append( 'nonce', settingsImportJsVars.settings_import_file_nonce );
 
   jQuery.ajax({
-    url: efmlImportJsVars.ajax_url,
+    url: settingsImportJsVars.ajax_url,
     type: "POST",
     data: request,
     contentType: false,