+++ app/ExternalFiles/Forms.php
@@ -222,6 +222,12 @@ class Forms {
 						'variant' => 'secondary',
 						'text'    => __( 'Cancel', 'external-files-in-media-library' ),
 					),
+					array(
+						'action'  => 'window.open("' . Helper::get_plugin_support_url() . '", "_blank" )',
+						'className' => 'efml-help-button',
+						'variant' => 'secondary',
+						'text'    => __( 'Need help?', 'external-files-in-media-library' ),
+					),
 				),
 			);
 
@@ -289,6 +295,12 @@ class Forms {
 					'action'  => 'closeDialog();',
 					'variant' => 'secondary',
 					'text'    => __( 'Cancel', 'external-files-in-media-library' ),
+				),
+				array(
+					'action'  => 'window.open("' . Helper::get_plugin_support_url() . '", "_blank" )',
+					'className' => 'efml-help-button',
+					'variant' => 'secondary',
+					'text'    => __( 'Need help?', 'external-files-in-media-library' ),
 				),
 			),
 		);