+++ changelog.md
@@ -11,6 +11,7 @@
 
 - Optimized the settings objects for future independence from this plugin
 - Usage of external embedded YouTube URLs in Block Editor is now possible
+- Choose your YouTube videos from local media library in Elementors Video widget
 
 ### Fixed