+++ changelog.md
@@ -8,6 +8,7 @@
 - Added more hooks
 - Added Third Party support for Block Editor and Elementor
 - Added link to support forum on our plugin in plugin list
+- Added link to support forum in add URLs dialog
 
 ### Optimized